requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.5
django-cleanup==5.2.0
django-crispy-forms==1.12.0
django-heroku==0.3.1
django-taggit==1.5.1
gunicorn==20.1.0
Markdown==3.3.4
Pillow==8.3.1
psycopg2==2.9.1
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.3.0
